Fix crash when saving field "Requirements Files".
The code tried to load the file contents into a list. However, the file may not even exist at this time,
and the list was subsequently never used, so we might as well not read the file right now.
When we do attempt to read it, fail if the file does not exist.